LayoutTests/js/dom/function-bind-expected.txt

Tests Function.bind.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new (decodeURI.bind())() threw exception TypeError: function is not a constructor.
PASS (new (String.bind())('foo')).toString() is 'foo'
PASS result is "[object Window] -> x:1, y:2"
PASS result is "'a' -> x:'b', y:1"
PASS result is "'a' -> x:'b', y:'c'"
PASS result is "[object Object] -> x:1, y:2"
PASS result is "[object Object] -> x:'b', y:1"
PASS result is "[object Object] -> x:'b', y:'c'"
PASS f instanceof F is true
PASS f instanceof G is true
PASS f instanceof H is true
PASS g instanceof F is true
PASS g instanceof G is true
PASS g instanceof H is true
PASS h instanceof F is true
PASS h instanceof G is true
PASS h instanceof H is true
PASS "prototype" in F is true
PASS "prototype" in G is false
PASS "prototype" in H is false
PASS Function.bind.call(undefined) threw exception TypeError: Cannot bind non-function object..
PASS abcAt(1) is "b"
PASS new abcAt(1) threw exception TypeError: function is not a constructor.
PASS boundFunctionPrototypeAccessed is false
PASS Function.bind.length is 1
PASS successfullyParsed is true

TEST COMPLETE
